// client/src/context/actions.ts
enum ActionTypes {
    DISPLAY_ALERT = 'DISPLAY_ALERT',
    CLEAR_ALERT = 'CLEAR_ALERT',

    AUTH_BEGIN = 'AUTH_BEGIN',
    AUTH_SUCCESS = 'AUTH_SUCCESS',
    AUTH_ERROR = 'AUTH_ERROR',

    UPDATE_USER_BEGIN = 'UPDATE_USER_BEGIN',
    UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS',
    UPDATE_USER_ERROR = 'UPDATE_USER_ERROR',

    LOGOUT_USER = 'LOGOUT_USER',

    TOGGLE_SIDEBAR = 'TOGGLE_SIDEBAR',

    HANDLE_CHANGE = 'HANDLE_CHANGE',

    CLEAR_VALUES = 'CLEAR_VALUES',

    CREATE_JOB_BEGIN = 'CREATE_JOB_BEGIN',
    CREATE_JOB_SUCCESS = 'CREATE_JOB_SUCCESS',
    CREATE_JOB_ERROR = 'CREATE_JOB_ERROR',

    GET_JOBS_BEGIN = 'GET_JOBS_BEGIN',
    GET_JOBS_SUCCESS = 'GET_JOBS_SUCCESS',

    SET_EDIT_JOB = 'SET_EDIT_JOB',

    DELETE_JOB_BEGIN = 'DELETE_JOB_BEGIN',

    EDIT_JOB_BEGIN = 'EDIT_JOB_BEGIN',
    EDIT_JOB_SUCCESS = 'EDIT_JOB_SUCCESS',
    EDIT_JOB_ERROR = 'EDIT_JOB_ERROR',

    SHOW_STATS_SUCCESS = 'SHOW_STATS_SUCCESS',
    SHOW_STATS_BEGIN = 'SHOW_STATS_BEGIN',

    CLEAR_FILTERS = 'CLEAR_FILTERS',

    CHANGE_PAGE = 'CHANGE_PAGE',
}

export default ActionTypes